Serialization of module state
Unifina-core now supports the serialization of RunningSignalPaths. In other
words, an associated SignalPath and its object graph will be serialized to
database every minute or so. If the running signal path is stopped or the
thread terminates the running signal path can be resumed by deserializing the
saved database state.

Serialization was implemented using fst-serialization library.