Fix threading issue in MessageHubSpec
There is a possible scenario in which MessageHub processes the message before
the test invokes wait(), which causes the test to wait for 500ms on a monitor
that will never be notified. Not a serious issue in this case, fortunately, but
it is more correct to send the message inside the synchronize block thereby
avoiding this situation.