Added removed item output to RemoveFromMap module, fixed bugs
When testing the added feature, noticed that tests run the modifying operation four times to same input map instance! This could be trouble if e.g. there are two modules that get the same output; one modifies the output, the other sees the modified map. To fix this, all Maps that pass though Output are wrapped in UnmodifiedMap.