CORE-621 code review fixes: flushing should now be much thread-safer
when user is thrown away, it's double checked for just-added metrics; when counter is zeroed, same thing.

There's a sync on "stats" now that could be done without, but it's not on the "critical path" (normal increment). Without sync, following scenario is possible:
* threads 1...N add new user to stats (line 45)
* thread N+1 removes user (line 91), checks counters, finds nothing, and finishes flush()

Now there's N threads incrementing counters that have been checked to be zero earlier and forgotten. This is of course not going to common, since deleting user happens only after one full reporting interval without increments.