Fix SendToStream bug (CORE-730)
Saving a canvas failed when module SendToStream appeared on canvas and its
parameter "stream" was connected to another module's output. For some reason,
this resulted in a Stream object being present in configuration JSON when
attempting to save canvas. I fixed this by enforcing that a String value is
always be present in configuration.