CORE-833: separate serialization byte to own table
Fetching a Canvas from database can be very expensive if its serialization is
large (e.g tens of megabytes.) This is especially problematic when fetching
lists of Canvases.

Unfortunately Grails only offers lazy fetching for associations and not within
a domain class / table. So I made a separate Serialization domain class that is
associated with (and dependent on) a Canvas.

Not the best possible solution to