CORE-912: move webpack build hook to PackagingEnd
Before, webpack build hook was executed on the CompileStart event, i.e., when
compilation of grails/java resources began. The problem with this approach is
that webpack compilation is activated also when only unit and/or integration
tests are run. This is especially obtrusive when running and developing tests
on the command line in daemon mode.

I moved the webpack hook to PackagingEnd, which is called at the end of
packaging (prior to Tomcat server being started and after web.xml is
generated). This means that the hook is executed only when the server is
started and not in unit/integration test runs.

Hopefully, this change doesn't cause other problems.