Core 894 account keys (#257)
* added (invisible) recaptcha on signup page

* added webpack and started to build accountkeyhandler in react

* merge conflict

* react stuff profilePage

* accounthandler

* more react-bootstrap components

* flowtype added

* continued tests

* accountSelector should work now

* package.json update

* updated README

* renamed account -> integration key

* add webpack-dev-server to package.json

* commonsChunkPlugin added

* bug fixes

* fix ethereum integration service bugs

* hook to build js

* recaptcha invisible -> v2

* CORE-909: migrate etherum keys as integration keys

Modules `EthereumCall` and `SolidityModule` have Ethereum public addresses
and private keys as module settings. This commit contains a migration script
that migrates them to Streamr as Integration Keys.

* CORE-909: EthereumAccountParameter

Replaces Ethereum public address and private key configurations in module
options of modules EthereumCall and SolidityModule. From now on an existing
integration key (of service ETHEREUM) is chosen from a drop-down parameter
instead.

* fixes by the pr review

* CORE-909 Ethereum account parameter (#260)

* CORE-909: migrate etherum keys as integration keys

Modules `EthereumCall` and `SolidityModule` have Ethereum public addresses
and private keys as module settings. This commit contains a migration script
that migrates them to Streamr as Integration Keys.

* CORE-909: EthereumAccountParameter

Replaces Ethereum public address and private key configurations in module
options of modules EthereumCall and SolidityModule. From now on an existing
integration key (of service ETHEREUM) is chosen from a drop-down parameter
instead.

* captcha

* typo

* fixed many things according to comments

* test fixed

* unit test for EthereumAccountParameter

* CORE-912: move webpack build hook to PackagingEnd

Before, webpack build hook was executed on the CompileStart event, i.e., when
compilation of grails/java resources began. The problem with this approach is
that webpack compilation is activated also when only unit and/or integration
tests are run. This is especially obtrusive when running and developing tests
on the command line in daemon mode.

I moved the webpack hook to PackagingEnd, which is called at the end of
packaging (prior to Tomcat server being started and after web.xml is
generated). This means that the hook is executed only when the server is
started and not in unit/integration test runs.

Hopefully, this change doesn't cause other problems.