CORE-1014: Change TnC and Privacy Policy to mention AG
TODO: check that the Privacy Police indeed conforms to https://www.admin.ch/opc/en/classified-compilation/19920153/index.html#id-2

Section 2 Art 11a paragraph 3 seems to not apply to Streamr or Streamr personnel, since we don't regularly process sensitive personal data. It's not part of normal operation.

Art 8 specifically states "The information must normally be provided in writing, in the form of a printout or a photocopy, and is free of charge." so removed the "reasonable costs" part.