CORE-840 canvas no longer gets saved on load
Reconstructing a canvas on load may make a change to its json, in which case the canvas will get auto-saved by Hibernate. Added an explicit discard statement. This could not be unit tested, because the mock GORM is not consistent with real GORM and apparently doesn't have the auto-save feature in Grails 2.3.x. This might be fixed in 2.5.x.