CanvasDeleteTask created for adhoc canvas at end
Method SignalPathRunner#destroy attempts to create CanvasDeleteTask(s)
via SignalPathService via CanvasService. But because the adhoc canvases
in question are never marked as stopped, the attempt fails and causes an
exception to be thrown instead.

Fixed it so that adhoc canvases are marked as stopped and that
CanvasService allows creation of delayed delete tasks even if canvas is
running.