CORE-1384: test subscription case-insensitivity (#450)
I was worried that case-sensitivity issues could cause problems when matching the Ethereum addresses of `PaidSubscription` to those of `IntegrationKey`. Then I (re-)learned that MySQL is by default case-insensitive with respect to strings. Therefore string matching in `WHERE` clause and uniqueness constraints are case-insensitive so no problems arise!

Anyway I thought it would be a good idea to encode this assumptions as tests just in case there were any changes to our environment in the future and this property would no longer hold.