Access realtime/historical data via API instead of directly from Redis/Cassandra (#650)
This PR makes `engine-and-editor` access stream content via the API instead of directly. This is necessary for decoupling the Engine from the Network implementation in preparation for the Corea network release.

- Canvases now subscribe to data via the API instead of accessing Redis/Cassandra directly for realtime/historical runs respectively.
- Canvases now publish data via the API instead of producing to Kafka directly.
- To achieve the above, a shared `StreamrClient` instance is available to the `DataSource` and canvas modules via `Globals`. The `StreamrClient` authenticates as the user who started the canvas.
- Remove all references to Kafka from the codebase.
- `RunCanvasSpec` couldn't work anymore, as the API is not up in integration tests. Ported the same test content to an end-to-end JS test.

**Also fixed some loosely related stuff that I encountered while working on the above, sorry for the noisy PR:**

- [CORE-1781](https://streamr.atlassian.net/browse/CORE-1781): Add partition selection support to `SendToStream` module.
- Throttle the rate at which messages are published during a csv upload.
- Allow stream partition count to be updated via the API.
- `PUT` requests to stream endpoint now behave the same as other PUT endpoints: return `200` and render the updated resource (was: return 204 and empty response).
- Other minor codebase maintenance and README updates.

**Out of scope for this PR:**

- There are still some direct usage of Cassandra in `StreamService` related to deleting messages. These can't be removed yet, as the P2P network API doesn't support deleting messages yet. In the Corea network we'll initially have just one centralized Cassandra cluster for storage, so the existing functionality in `engine-and-editor` will work, and it is not blocking the Corea network from being deployed. The delete support is outlined in [NET-57](https://streamr.atlassian.net/browse/NET-57).

**Notable shortcomings/hacks:**

In order to access data via the API, the connection needs to be authenticated. So the canvas needs to authenticate as the user who started it. For this purpose, the running canvas will need to use some credential of the user to authenticate the session. Currently an arbitrarily selected API key is used (all users have at least one API key for some other, earlier reason, but now this requirement is enforced by this implementation).

This approach may have some problems in various corner cases. Such as when an API key is revoked. If that key happens to be used by a running canvas, and the session token created when the canvas was started happens to expire, then the API key is used again to start a new session. For a revoked API key, this procedure will fail, and the canvas will no longer be able to interact with the API.