BACK-67: Create Stream ID with Sandbox name (#850)
Allow users to define explicit ID when they create a new stream. The stream ID must start with "sandbox" and have a some path appended to that. E.g. "sandbox/foo/bar" is a valid stream name.

Stream id is not a required field: if we don't specify an ID, the backend generates a random ID.

Added support for encoded slashes in API endpoints:
- embedded tomcat plugin allows encoded slash in URI when using grails run-app and grails run-war
- all end-to-end tests use the API via Nginx